Better error handling in magics for bad magic arguments
